// src/app/api/model/cu-store-model.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CUStoreModel { 
    storeId?: string | null;
    storeCode: string;
    storeName: string;
    address?: string | null;
    phone?: string | null;
    openedOn?: string | null;
    isActive?: boolean;
}